feat(http,ffi): Wire backup, block, and encrypted index into CLI/HTTP; align ACP/NAC stores
Extract shared backup and block-verify logic from FFI into the db crate
so both FFI and CLI/HTTP share the same algorithms. Wire BackupOperations,
BlockOperations, and EncryptedIndexOperations into the HTTP server and CLI.
Align FFI ACP/NAC stores with CLI (PersistentAcpStore and
PersistentZanzibarStore for file-based storage).
