+++ crates/ffi/src/document/mod.rs
@@ -2,5 +2,4 @@ mod create;
 mod parse;
 
 pub use create::collection_create;
-pub(crate) use create::json_to_graphql_input;
 pub use parse::{is_json_array, parse_duration, parse_string_array};
